lab6

+++ mysite/polls/views.py
@@ -1,10 +1,16 @@
 from django.http import HttpResponse
-from django.shortcuts import render
+from django.shortcuts import render, get_object_or_404
 from rest_framework import status
-from rest_framework.decorators import api_view
+from rest_framework.authentication import SessionAuthentication, BasicAuthentication, TokenAuthentication
+from rest_framework.decorators import api_view, authentication_classes, permission_classes
+from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 from .models import Osoba, Stanowisko
 from .serializers import OsobaModelSerializer, StanowiskoModelSerializer
+
+
+class BearerTokenAuthentication(TokenAuthentication):
+    keyword = u"Bearer"
 
 
 def index(request):
@@ -12,23 +18,40 @@ def index(request):
 
 
 @api_view(['GET'])
+@authentication_classes([SessionAuthentication, BasicAuthentication])
+@permission_classes([IsAuthenticated])
 def osoba_list(request):
     """
     Lista wszystkich obiektów modelu Osoba
     """
     if request.method == 'GET':
-        persons = Osoba.objects.all()
+        persons = Osoba.objects.filter(wlasciciel=request.user)
         serializer = OsobaModelSerializer(persons, many=True)
         return Response(serializer.data)
 
 
-@api_view(['GET', 'PUT', 'DELETE'])
+@api_view(['GET'])
 def osoba_detail(request, pk):
 
     try:
         osoba = Osoba.objects.get(pk=pk)
     except Osoba.DoesNotExist:
         return Response(status=status.HTTP_404_NOT_FOUND)
+
+    if request.method == 'GET':
+        serializer = OsobaModelSerializer(osoba)
+        return Response(serializer.data)
+
+
+@api_view(['PUT', 'GET'])
+@authentication_classes([SessionAuthentication, BasicAuthentication])
+@permission_classes([IsAuthenticated])
+def osoba_update(request, pk):
+
+    osoba = get_object_or_404(Osoba, pk=pk)
+
+    if osoba.wlasciciel != request.user:
+        return Response(status=status.HTTP_401_UNAUTHORIZED)
 
     if request.method == 'GET':
         serializer = OsobaModelSerializer(osoba)
@@ -41,19 +64,46 @@ def osoba_detail(request, pk):
             return Response(serializer.data, status=status.HTTP_202_ACCEPTED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
+
+@api_view(['DELETE', 'GET'])
+@authentication_classes([BearerTokenAuthentication])
+@permission_classes([IsAuthenticated])
+def osoba_delete(request, pk):
+
+    osoba = get_object_or_404(Osoba, pk=pk)
+
+    if osoba.wlasciciel != request.user:
+        return Response(status=status.HTTP_401_UNAUTHORIZED)
+
+    if request.method == 'GET':
+        serializer = OsobaModelSerializer(osoba)
+        return Response(serializer.data)
+
     if request.method == 'DELETE':
         osoba.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
 
 @api_view(['POST'])
+@authentication_classes([SessionAuthentication, BasicAuthentication])
+@permission_classes([IsAuthenticated])
 def osoba_add(request):
     if request.method == 'POST':
         serializer = OsobaModelSerializer(data=request.data)
         if serializer.is_valid():
-            serializer.save()
+            serializer.save(wlasciciel=request.user)
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+
+@api_view(['GET'])
+@authentication_classes([BearerTokenAuthentication])
+@permission_classes([IsAuthenticated])
+def osoba_stanowisko(request, id):
+    if request.method == 'GET':
+        osoby = Osoba.objects.filter(stanowisko=id)
+        serializer = OsobaModelSerializer(osoby, many=True)
+        return Response(serializer.data)
 
 
 @api_view(['GET'])
